Refactor (#4)
* Separate code for better reading

* Balance attack speed
